crmdig prefisso + gestione traduzioni soggetti
The pair_subject_object function in main_object_demo.py implemented an ex post cleaning mechanism to eliminate the association between the IRI of an appellation and all literal values extracted from the same "Soggetti" cell value using separators. In practice, it corrected an ‘excessive’ generation mechanism  (due to the issue with conditional mappings). However, by adding the translation, this function was no longer able to manage the cleaning mechanism.

It was updated to handle appellations with translations (e.g. “Pesce spada [Swordfish]”).
What it does now:

1 - Reads the Subjects column from the CSV files and separates multiple values (“;”).

2- It automatically recognises translation delimiters (e.g. ‘[ ]’, ‘< >’, ‘( )’) and builds a dictionary that associates each normalised form with the pair {original form, translation}.

3 - When checking triples with crm:P190_has_symbolic_content, if the speaking part of the IRI does not match the literal, it checks whether the literal corresponds to the recorded translation.

4- If so, the triple is retained; if not, it is removed.

In this way, pairs such as ‘feto’ == ‘fetus’ or “pesce_spada” == ‘swordfish’ are recognised as valid translations and are not deleted.

IN MAPPING AND UDF: update of function: idlab-fn:multiple_separator_split_explode, so as to manage the possibility of considering either the translation inside the characters of the original text or outside of it.